+++ .cargo/config.toml
@@ -1,0 +1,2 @@
+[alias]
+rpg = "run -q --bin rust-project-goals-cli"
