+++ crates/mdbook-goals/Cargo.toml
@@ -18,10 +18,8 @@ rust_team_data = { git = "https://github.com/rust-lang/team" }
 lazy_static = "1.5.0"
 progress_bar = "1.0.5"
 chrono = "0.4.38"
-aws-config = "1.5.8"
-aws-sdk-bedrock = "1.57.0"
-aws-sdk-bedrockruntime = "1.55.0"
 tokio = { version = "1.41.0", features = ["full"] }
 handlebars = { version = "6.1.0", features = ["dir_source"] }
 comrak = "0.29.0"
 rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
+rust-project-goals-llm = { version = "0.1.0", path = "../rust-project-goals-llm" }
