+++ crates/mdbook-goals/src/mdbook_preprocessor.rs
@@ -10,8 +10,7 @@ use mdbook::BookItem;
 use regex::{Captures, Regex};
 use rust_project_goals::util::GithubUserInfo;
 
-use crate::goal::{self, format_team_asks, GoalDocument, Status, TeamAsk};
-use crate::{re, team};
+use rust_project_goals::{re, team, goal::{self, format_team_asks, GoalDocument, Status, TeamAsk}};
 
 const LINKS: &str = "links";
 const LINKIFIERS: &str = "linkifiers";
