+++ crates/rust-project-goals-cli/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "rust-project-goals-cli"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
