+++ crates/rust-project-goals-llm/src/lib.rs
