+++ crates/rust-project-goals/Cargo.toml
@@ -5,7 +5,12 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.94"
+chrono = "0.4.39"
 disk-persist = "0.1.0"
+lazy_static = "1.5.0"
+regex = "1.11.1"
 reqwest = "0.12.9"
 serde = "1.0.216"
+serde_json = "1.0.133"
 walkdir = "2.5.0"
+rust_team_data = { git = "https://github.com/rust-lang/team" }
