+++ crates/rust-project-goals/src/lib.rs
@@ -1,2 +1,7 @@
 pub mod util;
 pub mod markwaydown;
+pub mod gh;
+pub mod json;
+pub mod re;
+pub mod team;
+pub mod goal;
